pyproject.toml
### [project] section: 
###     see https://peps.python.org/pep-0621/
### [build-system] section: 
###     see https://peps.python.org/pep-0517/
###     see https://peps.python.org/pep-0518/
### see https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "the-cradle-colorlight-i9-ecp5-amaranth-hdl-by-sporniket"
version = "0.1.0"
description = "An implementation of \"the cradle\" for the ECP5-based colorlight i9 written with the Amaranth HDL"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["amaranth","amaranth-hdl","nmigen", "hdl", "fpga", "ecp5", "colorlight i9"]
license = {file = "LICENSE"}
authors = [{name="Sporniket", email="sporniket.studio@gmail.com"}]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
]
dependencies = [
    "amaranth-stuff-by-sporniket @ git+https://github.com/sporniket/amaranth-stuff@v0.0.6",
    'importlib-metadata; python_version>"3.8"',
]

[project.urls]
homepage = "https://github.com/sporniket/the-cradle-colorlight-i9-ecp5-amaranth-hdl"
#TODO documentation = "https://readthedocs.org"
repository = "https://github.com/sporniket/the-cradle-colorlight-i9-ecp5-amaranth-hdl.git"
bug-tracker = "https://github.com/sporniket/the-cradle-colorlight-i9-ecp5-amaranth-hdl/issues"

[tool.pdm.scripts]
_.env_file = ".env.toolchain"
# --- bleeding edge dependencies update ---
display-dependency-updates = "python3 display-dependency-updates.py"
# --- cleaning
_clean_tons_of_files = "python3 reclean.py"
clean = { composite = [ "_clean_tons_of_files" ] }
# --- linting ---
reformat = "python3 -m black ."
_lint_errors = "python3 -m flake8 src --count --select=E9,F63,F7,F82 --show-source --statistics"
_lint_warnings = "python3 -m flake8 src --count --ignore=E266 --exit-zero --max-complexity=10 --max-line-length=127 --statistics"
lint_report = { composite = [ "_lint_errors", "_lint_warnings" ] }
lint_ci = { composite = [ "reformat", "_lint_errors" ] }
lint = { composite = [ "reformat", "lint_report" ] }
# --- building ---
make_bitstream = "python3 -m the_cradle.builder"
make_bitstream_vga = "python3 -m the_cradle.builder vga"
make_bitstream_pal = "python3 -m the_cradle.builder pal"
make_bitstream_quart_hd = "python3 -m the_cradle.builder quart_hd"
deploy_with_openfpgaloader = "openFPGALoader -c cmsisdap -m ./build/top.bit"
make_and_deploy_ofl = { composite = ["clean", "make_bitstream", "deploy_with_openfpgaloader"] }
make_and_deploy_ofl_vga = { composite = ["clean", "make_bitstream_vga", "deploy_with_openfpgaloader"] }
make_and_deploy_ofl_pal = { composite = ["clean", "make_bitstream_pal", "deploy_with_openfpgaloader"] }
make_and_deploy_ofl_quart_hd = { composite = ["clean", "make_bitstream_quart_hd", "deploy_with_openfpgaloader"] }
# the default 'make and deploy' will use openFPGALoader.
make_and_deploy = { composite = ["make_and_deploy_ofl"] }
make_and_deploy_vga = { composite = ["make_and_deploy_ofl_vga"] }
make_and_deploy_pal = { composite = ["make_and_deploy_ofl_pal"] }
make_and_deploy_quart_hd = { composite = ["make_and_deploy_ofl_quart_hd"] }
# --- test & coverage ---
_coverage_run = "python3 -m coverage run --source=ecp5 --branch -m pytest"
_coverage_report = "python3 -m coverage report -m"
_coverage_html = "python3 -m coverage html"
_ci_only = { composite = ["_coverage_run","_coverage_report","_coverage_html"] }
ci = { composite = ["clean", "lint_ci", "_ci_only"] }

[tool.pdm.dev-dependencies]
lint = [
    "black>=23.9.1",
    "flake8>=6.1.0",
]
ci = [
    "pytest>=7.4.2",
    "coverage>=7.3.2",
    "yowasp-yosys>=0.34.0.0.post591",
]
scripting = [
    "rich>=13.6.0",
]
build = [
    "pdm>=2.9.3",
    "pytest>=7.4.2",
    "twine>=4.0.2",
]
ecp5 = [
    "yowasp-nextpnr-ecp5>=0.6.0.0.post445",
]
